requirements.txt
asgiref==3.2.10
Django==3.1.2
django-adminlte-3==0.1.6
django-cors-headers==3.5.0
django-mysql==3.10.0
djangorestframework==3.12.1
export==0.1.2
import==0.0.1
mysql==0.0.2
mysqlclient==2.0.1
Pillow==8.0.1
python-http-client==3.3.1
pytz==2020.1
sendgrid==3.6.5
sendgrid-django==4.2.0
sqlparse==0.4.1
